5.Animation Loop
The loop function is responsible for the animation and drawing on the canvas.

It uses window.requestAnimFrame to ensure smooth animation by synchronizing with the browser's repaint cycle.

Within the loop function, it clears the canvas, calls the draw function to update and draw the tentacles, and repeats the loop.